Move crawler report classes into model.
