+++ crawler/crawler/eq.py
@@ -2,28 +2,11 @@ import json
 import math
 import time
 from datetime import datetime
-from typing import Literal, TypedDict
+from typing import TypedDict
 
-from sqlmodel import SQLModel
+from model import EqReport, Geometry, Intensity, Point
 
 from .crawler import Crawler
-
-Point = tuple[float, float]
-Intensity = tuple[int, int, int]
-
-
-class Geometry(TypedDict):
-    type: Literal["Point"]
-    coordinates: Point
-
-
-class EqReport(SQLModel):
-    timestamp: int
-    geometry: Geometry
-    scale: float
-    intensity: Intensity
-    link: str
-    img: str
 
 
 class EqQuery(TypedDict):
